+++ logs.js
@@ -1,4 +1,4 @@
-const logs = {
+export const logs = {
   start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
   end: [
       'Результат удара [playerWins]: [playerLose] - труп',
